Reordering file structure and renaming files for better navigation and readbility.
