+++ src/structure/genericdouble/DoubleLinkedListInterface.java
