+++ src/structure/genericdouble/DoublyLinkedList.java
