+++ src/structure/genericdouble/DoublyLinkedListTest.java
