+++ src/structure/genericsingle/SingleLinkedList.java
