+++ src/structure/genericsingle/SingleLinkedListInterface.java
