+++ src/structure/genericsingle/SingleLinkedListTest.java
@@ -1,0 +1,26 @@
+package src.structure.genericsingle;
+
+public class SingleLinkedListTest {
+    
+    private SingleLinkedList<Integer> list = new SingleLinkedList<Integer>();
+
+    private static final Integer ELEMENT_A = 1;
+	private static final Integer ELEMENT_B = 2;
+	private static final Integer ELEMENT_C = 3;
+	private static final Integer ELEMENT_D = 4;
+	private static final Integer ELEMENT_X = -1;
+	private static final Integer ELEMENT_Z = -2;
+
+    /***** 0-ELEMENT LINKED LIST TESTS ******/
+
+    /***** SINGLE LINKED LIST TESTS ******/
+
+    /***** TWO ELEMENT LINKED LIST TESTS ******/
+
+    /***** THRE ELEMENT SINGLE LINKED LIST TESTS ******/
+
+
+   
+
+
+}
