+++ src/structure/genericsingle/Test.java
@@ -1,0 +1,5 @@
+package src.structure.genericsingle;
+
+public @interface Test {
+
+}
