+++ src/structure/genericstack/Stack.java
