+++ src/structure/genericstack/StackInterface.java
@@ -1,4 +1,4 @@
-package src.structure.interfaces;
+package src.structure.genericstack;
 
 /**
  * @author Tenneh Kanneh
@@ -8,7 +8,7 @@ package src.structure.interfaces;
  *
  * @param <T> the type of elements held in this stack
  */
-public interface GenericStackInterface<T> {
+public interface StackInterface<T> {
     
     /**
      * Adds the specified element to the top of the stack.
