+++ src/structure/genericstack/StackTest.java
@@ -1,0 +1,5 @@
+package src.structure.genericstack;
+
+public class StackTest {
+    // Stack for undo
+}
