--- src/structure/test/GenericStackTest.java
@@ -1,5 +1,0 @@
-package src.structure.test;
-
-public class GenericStackTest {
-    // Stack for undo
-}
